requirements.txt
flask==3.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
huggingface_hub==0.16.4
python-docx==0.8.11
gunicorn==21.2.0
tqdm==4.66.1
